requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.6.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dacite==1.9.2
dask==2025.5.0
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
dill==0.3.9
dm-tree==0.1.8
docstring_parser==0.16
etils==1.12.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.68.0
graphviz==0.20.3
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
ImageHash==4.3.1
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
immutabledict==4.2.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.32.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==6.0.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_code_formatter==3.0.2
jupyterlab_execute_time==3.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.9
keras==3.8.0
keras-hub==0.19.0
keras-nlp==0.19.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
memoized-property==1.0.3
mistune==3.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multimethod==1.12
murmurhash==1.0.13
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbresult==0.1.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.6
nvidia-nccl-cu12==2.26.5
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
pmdarima==2.0.4
preshed==3.0.10
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pylint==3.3.4
pyparsing==3.2.1
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
sacremoses==0.1.1
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.16.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
torch @ https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.6.0%2Bcpu.cxx11.abi-cp312-cp312-linux_x86_64.whl#sha256=a45917579548be5c7c97c04531161aa5e3b100e37b815bfdf54402fc0b4a4f36
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typeguard==4.4.2
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.3
visions==0.7.6
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.1.4
ydata-profiling==4.12.2
zipp==3.21.0
